Cart - updated to use RQ provider and fixed bug with provider
