--- src/app/providers/providers.tsx
@@ -1,35 +1,0 @@
-'use client'
-
-import React from 'react'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
-import { ReactQueryStreamedHydration } from '@tanstack/react-query-next-experimental'
-
-import { Toaster } from '@/components/ui/sonner'
-
-interface ProviderProps {
-	children: React.ReactNode
-}
-
-export default function Providers({ children }: ProviderProps) {
-	const [queryClient] = React.useState(
-		() =>
-			new QueryClient({
-				defaultOptions: {
-					queries: {
-						staleTime: 1000 * 60 * 60 * 1,
-						refetchOnReconnect: false,
-						refetchOnMount: false,
-					},
-				},
-			}),
-	)
-
-	return (
-		<QueryClientProvider client={queryClient}>
-			<ReactQueryStreamedHydration>{children}</ReactQueryStreamedHydration>
-			<ReactQueryDevtools initialIsOpen={false} />
-			<Toaster />
-		</QueryClientProvider>
-	)
-}
