+++ src/components/layouts/navigation/nav-menu-link.tsx
@@ -1,5 +1,6 @@
 'use client'
 
+import { Route } from 'next'
 import Link from 'next/link'
 
 import {
@@ -16,7 +17,7 @@ interface NavMenuLinkProps {
 const NavMenuLink = ({ item }: NavMenuLinkProps) => {
 	return (
 		<NavigationMenuItem key={item.title}>
-			<Link href={item.href as string} legacyBehavior passHref>
+			<Link href={item.href as Route} legacyBehavior passHref>
 				<NavigationMenuLink className={navigationMenuTriggerStyle()}>
 					{item.title}
 				</NavigationMenuLink>
